Bluetooth/Bleak/ble-mqtt_bridge/requirements.txt
bleak==0.21.1
paho_mqtt==1.6.1
